MINIFICPP-1507 convert OutputStream::write to size_t

This closes #1083

+++ extensions/rocksdb-repos/RocksDbStream.h
@@ -77,7 +77,7 @@ class RocksDbStream : public io::BaseStream {
    * @param value value to write
    * @param size size of value
    */
-  int write(const uint8_t *value, int size) override;
+  size_t write(const uint8_t *value, size_t size) override;
 
  protected:
   std::string path_;
