+++ libminifi/include/io/FileStream.h
@@ -84,7 +84,7 @@ class FileStream : public io::BaseStream {
    * @param value value to write
    * @param size size of value
    */
-  int write(const uint8_t *value, int size) override;
+  size_t write(const uint8_t *value, size_t size) override;
 
  private:
   void seekToEndOfFile(const char* caller_error_msg);
