+++ libminifi/include/sitetosite/Peer.h
@@ -295,7 +295,7 @@ class SiteToSitePeer : public org::apache::nifi::minifi::io::BaseStream {
   using BaseStream::write;
   using BaseStream::read;
 
-  int write(const uint8_t* data, int len) override {
+  size_t write(const uint8_t* data, size_t len) override {
     return stream_->write(data, len);
   }
